requirements-test.txt
pytest-homeassistant-custom-component>=0.13
aioresponses>=0.7
pytest-asyncio>=0.23
